rework routing on admin side of things
