+++ src/app/admin/live/[slug]/layout.js
